on a tout fait en fait on est trop forts ceci est un message très serieux

+++ Service_DFT_Fract/Signal_4_12.h
@@ -3,7 +3,7 @@
 #define SIGNAL_H__
 
 
-extern float LeSignal[64];
+extern unsigned short int LeSignal[64];
 
 
 #endif
